Compiled the changes from Damia and Josia

+++ GroupProject_NutritionTracker/GroupProject/NutritionTracker/Models/HomeViewModel.cs
@@ -4,10 +4,10 @@
     {
         public DateTime SelectedDate { get; set; }
 
-        public int TotalCalories { get; set; }
-        public int TotalProtein { get; set; }
-        public int TotalCarbs { get; set; }
-        public int TotalFat { get; set; }
+        public double TotalCalories { get; set; }
+        public double TotalProtein { get; set; }
+        public double TotalCarbs { get; set; }
+        public double TotalFat { get; set; }
 
         // user's chosen max daily calorie limit
         public int MaxCalorieIntake { get; set; }
@@ -19,6 +19,6 @@
     {
         public int MealId { get; set; }
         public string Name { get; set; } = string.Empty;
-        public int TotalCalories { get; set; }
+        public double TotalCalories { get; set; }   // ← changed to double
     }
 }
